Refactor moving traversers into their own file.
